+++ docs/source/conf.py
@@ -1,6 +1,8 @@
 import os
 import sys
 
+from docs.source.versions import sphinx_versions
+
 sys.path.insert(0, os.path.abspath("."))
 sys.path.insert(0, os.path.abspath("../.."))
 import bigtree
@@ -53,15 +55,16 @@ language = "Python"
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = "sphinx_material"
-sphinx_versions = ["latest", "0.15.5", "0.14.8"]
 html_theme_options = {
     "nav_title": "bigtree Documentation",
     "base_url": "https://kayjan.github.io/bigtree",
-    "logo_icon": "&#xe88a",
-    "color_primary": "teal",
-    "color_accent": "amber",
     "repo_url": "https://github.com/kayjan/bigtree/",
     "repo_name": "bigtree",
+    "repo_type": "github",
+    "color_primary": "teal",
+    "color_accent": "amber",
+    "logo_icon": "&#xe88a",
+    "globaltoc_depth": 2,
     "heroes": {
         "index": "Tree Implementation and Methods for Python",
     },
@@ -74,6 +77,9 @@ html_show_sourcelink = True
 html_sidebars = {
     "**": [
         "globaltoc.html",
+        "localtoc.html",
+        "logo-text.html",
+        "searchbox.html",
         "node.html",
         "binarytree.html",
         "dag.html",