+++ docs/source/versions.py
@@ -1,0 +1,1 @@
+sphinx_versions = ["latest", "0.15.5", "0.14.8"]