+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.15.6] - 2023-01-20
 ### Added
 - DAGNode: Able to access and delete node children via name with square bracket accessor with `__getitem__` and `__delitem__` magic methods.
 - DAGNode: Able to delete all children for a node.
@@ -461,7 +463,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Utility Iterator: Tree traversal methods.
 - Workflow To Do App: Tree use case with to-do list implementation.
 
-[Unreleased]: https://github.com/kayjan/bigtree/compare/0.15.5...HEAD
+[Unreleased]: https://github.com/kayjan/bigtree/compare/0.15.6...HEAD
+[0.15.6]: https://github.com/kayjan/bigtree/compare/0.15.5...0.15.6
 [0.15.5]: https://github.com/kayjan/bigtree/compare/0.15.4...0.15.5
 [0.15.4]: https://github.com/kayjan/bigtree/compare/0.15.3...0.15.4
 [0.15.3]: https://github.com/kayjan/bigtree/compare/0.15.2...0.15.3